+++ Simulation_file.py
@@ -1,6 +1,6 @@
  # simulation file
 
-from LinkedList import *
+import LinkedList
 #from gui_experiment import *
 import Student
 
@@ -18,11 +18,48 @@ class simulation:
         run_sample,
     
     ):  
-
+        
+        # initialize students and slides
         self.students = Student.generate_students(17)
+        
+        self.sign = LinkedList.create_slides(20)
+
+    # def check_slide_visabilty(self, SlideNode, time_range, day):
+    #     if self.schedule[day] in range[time_range]:
+            
+    #         duration = time_range[1] - self.schedule[day] 
+            
+    #         self.view_slide(SlideNode, duration)
+
+    def check_slide_visabilty(self, SlideNode, time_range, day):
+    if self.schedule[day] in range[time_range]:
+        
+        duration = time_range[1] - self.schedule[day] 
+        
+        self.view_slide(SlideNode, duration)
+
+    
+    def simulate_sign_viewing(self):
+        time = 0
+        while time in range(simulation_span):
+            time = time + self.slide_duration
+            
+            
+            check_slide_visabilty(self.sign, time, 0)
+            self.sign
+
+
+
+
 
     def say_hello(self):
         print(self.students)
+
+
+    
+    
+    
+
 
 if __name__ == "__main__":
     sim = simulation(0,0,0,0,0,0,0,0,0)
